add Native settings implementation (#2346)
