+++ src/main/settings.ts
@@ -7,6 +7,7 @@
 import type { Settings } from "@api/Settings";
 import { IpcEvents } from "@shared/IpcEvents";
 import { SettingsStore } from "@shared/SettingsStore";
+import { mergeDefaults } from "@utils/mergeDefaults";
 import { ipcMain } from "electron";
 import { mkdirSync, readFileSync, writeFileSync } from "fs";
 
@@ -42,7 +43,22 @@ ipcMain.handle(IpcEvents.SET_SETTINGS, (_, data: Settings, pathToNotify?: string
     RendererSettings.setData(data, pathToNotify);
 });
 
-export const NativeSettings = new SettingsStore(readSettings("native", NATIVE_SETTINGS_FILE));
+export interface NativeSettings {
+    plugins: {
+        [plugin: string]: {
+            [setting: string]: any;
+        };
+    };
+}
+
+const DefaultNativeSettings: NativeSettings = {
+    plugins: {}
+};
+
+const nativeSettings = readSettings<NativeSettings>("native", NATIVE_SETTINGS_FILE);
+mergeDefaults(nativeSettings, DefaultNativeSettings);
+
+export const NativeSettings = new SettingsStore(nativeSettings);
 
 NativeSettings.addGlobalChangeListener(() => {
     try {
